05-scaling-reliability/develop/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psutil==6.0.0
httpx==0.28.1
python-dotenv==1.0.1
